requirements.txt
pinecone-client[grpc]==2.2.1 
pinecone-datasets==0.5.0rc11
sentence-transformers==2.2.2
